refactor: reorganise tests
